digital_human/requirements.txt
Flask-Sockets
gevent
gevent-websocket
pyjwt
markupsafe==2.0.1
Flask==1.1.4
Werkzeug==1.0.1